fixes for core/AmStatistics, apps/prometheus_exporter
rename AmStatisticsCounter.* -> AmStatistics.*
prettify code
avoid excess copying
fix prometheus_exporter thread name
fix segfault on empty prometheus_exporter cfg

+++ apps/prometheus_exporter/PrometheusExporter.h
@@ -18,7 +18,7 @@ class PrometheusExporter
     static PrometheusExporter* _instance;
 
     string ip;
-    int   port;
+    uint16_t port;
     string prefix;
     trsp_acl acl;
